Assignment3/ToDoProject/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
apturl==0.5.2
asgiref==3.7.2
astroid==3.0.2
asttokens==2.4.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.7.0
Brlapi==0.8.3
certifi==2023.11.17
chardet==5.2.0
click==8.1.7
colorama==0.4.6
command-not-found==0.3
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.9.0
distro-info==1.1+ubuntu0.1
Django==5.0.1
django-allauth==0.60.0
django-bootstrap4==23.4
dnspython==2.4.2
duplicity==2.1.5
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.19
flake8==7.0.0
future==0.18.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
ipython==8.19.0
isort==5.13.2
jedi==0.19.1
jeepney==0.8.0
keyring==24.3.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
lockfile==0.12.2
logilab-common==2.0.0
louis==3.20.0
macaroonbakery==1.3.4
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
monotonic==1.6
more-itertools==10.1.0
mypy-extensions==1.0.0
netifaces==0.11.0
numpy==1.26.3
oauthlib==3.2.2
olefile==0.47
openai==1.6.1
pandas==2.1.4
paramiko==3.4.0
parso==0.8.3
pexpect==4.9.0
Pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pur==7.3.1
pure-eval==0.2.2
pycairo==1.25.1
pycodestyle==2.11.1
pycups==2.0.1
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.2.0
Pygments==2.17.2
PyGObject==3.46.0
PyJWT==2.8.0
pylint==3.0.3
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.49
python3-openid==3.2.0
pytz==2023.3.post1
pyxdg==0.28
PyYAML==6.0.1
reportlab==4.0.8
requests==2.31.0
requests-oauthlib==1.3.1
SecretStorage==3.3.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
systemd-python==235
toml==0.10.2
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.1.0
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.13
wrapt==1.16.0
xdg==6.0.0
xkit==0.0.0
zipp==3.17.0
